+++ CHANGELOG.md
@@ -6,8 +6,7 @@ v3.0.0
   * Specification-arg-resolver can be used in GraalVM native images, but it requires several additional configuration steps. This is due to the fact that this library relies on Java reflection heavily. Please see [README_native_image.md](README_native_image.md) for the details
 * Modified Springdoc-openapi dependency to be compatible with spring boot 3.0
 * Refactored `IsMember`, `IsNotMember` specifications - they no longer use reflection explicitly. 
-* changed default join type in `@Join` from `INNER` to `LEFT`
-  I.e. if you have been using `@Join` with default join type:
+* changed default join type in `@Join` from `INNER` to `LEFT`. I.e. if you have been using `@Join` with default join type:
   ```java
   @Join(path="addresses", alias="a")
   ```